レビュー編集機能
ログイン時のみ投稿できるように設定
投稿者のみ編集できるように設定

+++ app/controllers/books_controller.rb
@@ -1,9 +1,11 @@
 class BooksController < ApplicationController
+  before_action :authenticate_user!, only: [:new, :create, :edit, :update, :destroy]
 
   def index
     @books = Book.all.order("created_at DESC").page(params[:page]).per(3)
     @review = Review.all
     @reviews = @review.includes(:book)
+    @user_reviews = current_user.reviews.order("created_at DESC") if current_user.present?
   end
 
   def new
@@ -16,7 +18,7 @@ class BooksController < ApplicationController
     if @book.save
       redirect_to root_path
     else
-      redirect_to new_book_path
+      redirect_to new_book_path, flash: {notice: '本の保存に失敗しました'}
     end
   end
   
@@ -28,23 +30,37 @@ class BooksController < ApplicationController
 
   def edit
     @book = Book.find(params[:id])
+    if @book.user_id == current_user.id
+      @book = Book.find(params[:id])
+    else
+      redirect_to action: 'show'
+    end
   end
 
   def update
     book = Book.find(params[:id])
     book.update(book_params)
-    redirect_to book_path(book.id)
+    if book.save
+      redirect_to book_path(book.id)
+    else
+      redirect_to action: 'update'
+    end
   end
 
   def destroy
     book = Book.find(params[:id])
-    book.destroy
-    redirect_to root_path
+    if book.user_id == current_user.id
+      book.destroy
+      redirect_to root_path
+    else
+      redirect_to action: 'show'
+    end
   end
 
   private
   def book_params
     params.require(:book).permit(:booktitle, :author, :bookimage).merge(user_id: current_user.id)
   end
+
 
 end
