+++ app/controllers/reviews_controller.rb
@@ -1,4 +1,6 @@
 class ReviewsController < ApplicationController
+  before_action :authenticate_user!, only: [:new, :create, :edit, :update, :destroy]
+
   def new
     @review = Review.new
   end
@@ -9,23 +11,40 @@ class ReviewsController < ApplicationController
 
   def create
     @review = Review.create(review_params)
-    redirect_to "/books/#{@review.book.id}"
+    if @review.save
+      redirect_to "/books/#{@review.book.id}"
+    else
+      redirect_to "/books/#{@review.book.id}", flash: {notice: 'レビューの保存に失敗しました'}
+    end
   end
 
   def edit
-    @review = Post.find(params[:id])
+    @review = Review.find(params[:id])
+    if @review.user_id == current_user.id
+      @review = Review.find(params[:id])
+    else
+      redirect_to "/books/#{@review.book_id}"
+    end
   end
 
   def update
-    review = Post.find(params[:id])
+    review = Review.find(params[:id])
     review.update(post_params)
-    redirect_to  new_book_review_path(review.id)
+    if review.save
+      redirect_to  new_book_review_path(review.id)
+    else
+      redirect_to action: 'update'
+    end
   end
 
   def destroy
-    review = Post.find(params[:id])
-    review.destroy
-    redirect_to root_path
+    review = Review.find(params[:id])
+    if review.user_id == current_user.id
+      review.destroy
+      redirect_to root_path
+    else
+      redirect_to "/books/#{review.book_id}"
+    end
   end
